Made KBDD script work on Linux
This seems to be a very specific way to get it to work using the properties of how
kbdd echoes to screen. This is not a general solution.

I am not sure why the general solution that works on OS X (the version of script
used is different from the one on Linux) does not work on Linux.

I have posted a question on StackOverflow
http://stackoverflow.com/questions/14554625/unix-script-command-not-capturing-stdin-when-launched-as-external-process

See #6.