Add miniSAT example
miniSAT is bit more unusual because the best way to interact with it is
through input and output files. Therefore, there is some intermediate
step to get a string into a file (e.g., like in a student submission)
and then we also need to capture the contents of the output file.

Closes #10.